backend/image/requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
Werkzeug==3.1.3
